+++ visualization/app/codeCharta/codeCharta.model.ts
@@ -192,9 +192,6 @@ export enum BlacklistType {
 export interface MarkedPackage {
 	path: string
 	color: string
-	attributes: {
-		[key: string]: any
-	}
 }
 
 export interface MetricData {